Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "appwrite>=11.0.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.13",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "pygeohash>=3.1.3",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "ultralytics>=8.3.158",
]

[dependency-groups]
dev = [
    "pytest-html>=4.1.1",
]
